+++ Plugin.BluetoothLE/AbstractGattCharacteristic.cs
@@ -61,15 +61,24 @@ namespace Plugin.BluetoothLE
                             .ToTask(cts.Token)
                             .ConfigureAwait(false);
 
-                        if (this.Value != buffer)
-                        {
-                            trans.Abort();
-                            throw new GattReliableWriteTransactionException("There was a mismatch response");
-                        }
+                        //if (this.Value != buffer)
+                        //{
+                        //    trans.Abort();
+                        //    throw new GattReliableWriteTransactionException("There was a mismatch response");
+                        //}
                         var seg = new BleWriteSegment(buffer, pos, len);
                         ob.OnNext(seg);
 
                         read = stream.Read(buffer, 0, buffer.Length);
+
+                        if (read > 0 && read < buffer.Length)
+                        {
+                            for (var index = read; index < buffer.Length; index++)
+                            {
+                                buffer[index] = 0;
+                            }
+                        }
+
                         pos += read;
                     }
                     await trans.Commit();