+++ Plugin.BluetoothLE/Extensions_Device.cs
@@ -47,6 +47,7 @@ namespace Plugin.BluetoothLE
                 .Select(x => x.ReadRssiContinuously(readInterval))
                 .Switch();
 
+
         /// <summary>
         /// Waits for connection to actually happen
         /// </summary>
@@ -99,7 +100,7 @@ namespace Plugin.BluetoothLE
                     .WhenConnected()
                     .Select(_ => device.WhenKnownCharacteristicsDiscovered(args.ServiceUuid, args.CharacteristicUuids))
                     .Switch()
-                    .Select(x => x.RegisterAndNotify(args.UseIndicateIfAvailable))
+                    .Select(x => x.RegisterAndNotify(args.UseIndicateIfAvailable, false))
                     .Switch()
                     .Subscribe(ob.OnNext);
 
@@ -116,7 +117,7 @@ namespace Plugin.BluetoothLE
 
 
         /// <summary>
-        /// Attempts to connect to the device, discover the characteristic and write to it
+        /// Discover the characteristic and write to it
         /// </summary>
         /// <param name="device"></param>
         /// <param name="serviceUuid"></param>
@@ -124,38 +125,31 @@ namespace Plugin.BluetoothLE
         /// <param name="data"></param>
         /// <returns></returns>
         public static IObservable<CharacteristicGattResult> WriteCharacteristic(this IDevice device, Guid serviceUuid, Guid characteristicUuid, byte[] data)
-        {
-            var obs = device
+            => device
                 .WhenKnownCharacteristicsDiscovered(serviceUuid, characteristicUuid)
-                .Select(x => x.Write(data))
+                .Select(x =>
+                {
+                    return x.Write(data);
+                })
                 .Switch();
-
-            device.ConnectIf();
-            return obs;
-        }
 
 
         /// <summary>
-        /// Attempts to connect to device, discover the characteristic and read it
+        /// Discover the characteristic and read it
         /// </summary>
         /// <param name="device"></param>
         /// <param name="serviceUuid"></param>
         /// <param name="characteristicUuid"></param>
         /// <returns></returns>
         public static IObservable<CharacteristicGattResult> ReadCharacteristic(this IDevice device, Guid serviceUuid, Guid characteristicUuid)
-        {
-            var obs = device
+            => device
                 .WhenKnownCharacteristicsDiscovered(serviceUuid, characteristicUuid)
                 .Select(ch => ch.Read())
                 .Switch();
 
-            device.ConnectIf();
-            return obs;
-        }
-
 
         /// <summary>
-        /// Will attempt to connect if necessary, discover the known characteristic, and read on a set interval
+        /// Discover the known characteristic, and read on a set interval
         /// </summary>
         /// <param name="device"></param>
         /// <param name="serviceUuid"></param>
@@ -163,15 +157,10 @@ namespace Plugin.BluetoothLE
         /// <param name="timeSpan"></param>
         /// <returns></returns>
         public static IObservable<CharacteristicGattResult> ReadIntervalCharacteristic(this IDevice device, Guid serviceUuid, Guid characteristicUuid, TimeSpan timeSpan)
-        {
-            var obs = device
+            => device
                 .WhenKnownCharacteristicsDiscovered(serviceUuid, characteristicUuid)
                 .Select(ch => ch.ReadInterval(timeSpan))
                 .Switch();
-
-            device.ConnectIf();
-            return obs;
-        }
 
 
         /// <summary>
@@ -184,8 +173,7 @@ namespace Plugin.BluetoothLE
         public static IObservable<IGattCharacteristic> GetKnownCharacteristics(this IDevice device, Guid serviceUuid, params Guid[] characteristicIds) =>
             device
                 .GetKnownService(serviceUuid)
-                .SelectMany(x => x.GetKnownCharacteristics(characteristicIds))
-                .Take(characteristicIds.Length);
+                .SelectMany(x => x.GetKnownCharacteristics(characteristicIds));
 
 
 
@@ -235,7 +223,10 @@ namespace Plugin.BluetoothLE
         public static IObservable<IGattCharacteristic> WhenKnownCharacteristicsDiscovered(this IDevice device, Guid serviceUuid, params Guid[] characteristicIds) =>
             device
                 .WhenConnected()
-                .SelectMany(x => x.GetKnownCharacteristics(serviceUuid, characteristicIds));
+                .SelectMany(x =>
+                {
+                    return x.GetKnownCharacteristics(serviceUuid, characteristicIds);
+                });
 
 
         /// <summary>